+++ R/initi_py.R
@@ -43,7 +43,7 @@ init_py <- function(py_ver = NULL, dgpsi_ver = NULL, reinstall = FALSE, uninstal
     ##For devel version
     dgpsi_ver <- c('cython>=0.29.30', 'dill>=0.3.2, <=0.3.5.1', 'jupyter>=1.0.0', 'matplotlib-base>=3.2.1', 'numba >=0.51.2',
                    'numpy >=1.18.2', 'pathos ==0.2.9', 'multiprocess ==0.70.13', 'psutil >=5.8.0', 'pybind11 >=2.10.0', 'pythran >=0.11.0',
-                   'scikit-build >=0.15.0', 'scikit-learn >=0.22.0', 'scipy >=1.4.1', 'tqdm >=4.50.2', 'tabulate >=0.8.7')
+                   'scikit-build >=0.15.0', 'scikit-learn >=0.22.0', 'scipy >=1.4.1', 'tqdm >=4.50.2', 'tabulate >=0.8.7', 'faiss-cpu >=1.7.4')
     env_name <- 'dgp_si_R_2_4_0_9000'
     ##For release version
     #dgpsi_ver <- 'dgpsi==2.4.0'
@@ -166,11 +166,17 @@ install_dgpsi <- function(env_name, py_ver, conda_path, dgpsi_ver, reinsatll = F
     }
   }
   if (Sys.info()[["sysname"]] == 'Linux' & !any(grepl("libstdc++.so.6.0.3",list.files("/usr/lib/x86_64-linux-gnu/"), fixed = TRUE))){
-    libstdc_path <- paste(gsub("bin.*$", "", conda_path), 'envs/', env_name, '/lib/libstdc++.so.6.0.3*', sep='')
-    system(paste("sudo cp", libstdc_path, "/usr/lib/x86_64-linux-gnu/"))
-    libstdc_sys_path <- "/usr/lib/x86_64-linux-gnu/libstdc++.so.6"
-    system(paste("sudo rm",libstdc_sys_path))
-    system(paste("sudo ln -s", "/usr/lib/x86_64-linux-gnu/libstdc++.so.6.0.3*", libstdc_sys_path))
+    cat("The required file 'libstdc++.so.6.0.30' or above is missing from /usr/lib/x86_64-linux-gnu/.")
+    ans <- readline(prompt="To proceed, would you like to grant sudo permissions to resolve the issue? (Y/N) ")
+    if ( tolower(ans)=='y'|tolower(ans)=='yes' ){
+      libstdc_path <- paste(gsub("bin.*$", "", conda_path), 'envs/', env_name, '/lib/libstdc++.so.6.0.3*', sep='')
+      system(paste("sudo cp", libstdc_path, "/usr/lib/x86_64-linux-gnu/"))
+      libstdc_sys_path <- "/usr/lib/x86_64-linux-gnu/libstdc++.so.6"
+      system(paste("sudo rm",libstdc_sys_path))
+      system(paste("sudo ln -s", "/usr/lib/x86_64-linux-gnu/libstdc++.so.6.0.3*", libstdc_sys_path))
+    } else {
+      stop("Please link /usr/lib/x86_64-linux-gnu/libstdc++.so.6 to 'libstdc++.so.6.0.30' or above and run init_py(reinstall = T).", call. = FALSE)
+    }
   }
   message("Installation finished. Please restart R.")
 }